Spostati alcuni metodi della gestione studenti e della gestione dei voti. Manuel Baldoni ha terminato la gestione degli studenti.
