+++ DataLayer/DL_ClassManagement.cs
@@ -1,6 +1,9 @@
 ﻿using SchoolGrades.DbClasses;
+using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Data.Common;
+using System.Data.SQLite;
 
 namespace SchoolGrades
 {
@@ -37,6 +40,40 @@ namespace SchoolGrades
                 cmd.Dispose();
             }
             return lc;
+        }
+
+        internal DataTable GetWeightedAveragesOfClass(Class Class,
+            string IdGradeType, string IdSchoolSubject, DateTime DateFrom, DateTime DateTo)
+        {
+            DataTable t;
+            using (DbConnection conn = Connect())
+            {
+                string query = "SELECT Grades.idGrade,Students.idStudent,lastName,firstName" +
+                ",SUM(weight)/100 AS 'GradesFraction', 1 - SUM(weight)/100 AS LeftToCloseAssesments" +
+                ",COUNT() AS 'GradesCount'" +
+                " FROM Classes_Students" +
+                " LEFT JOIN Grades ON Students.idStudent=Grades.idStudent" +
+                " JOIN Students ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass =" + Class.IdClass +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                " AND (Grades.idGradeType='" + IdGradeType + "'" +
+                " OR Grades.idGradeType IS NULL)" +
+                " AND Grades.idSchoolSubject='" + IdSchoolSubject + "'" +
+                " AND Grades.value IS NOT NULL AND Grades.value <> 0" +
+                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
+                " GROUP BY Students.idStudent" +
+                " ORDER BY GradesFraction ASC, lastName, firstName, Students.idStudent;";
+                // !!!! TODO change the query to include at first rows also those students that have no grades !!!! 
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
         }
     }
 }
