+++ DataLayer/DL_StudentManagement.cs
@@ -1,0 +1,352 @@
+﻿using SchoolGrades.DbClasses;
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.Common;
+using System.Data.SQLite;
+
+namespace SchoolGrades
+{
+    internal partial class DataLayer
+    {
+        internal int CreateStudentFromStringMatrix(string[,] StudentData, int? StudentRow)
+        {
+            // trova una chiave da assegnare al nuovo studente
+            int codiceStudente = NextKey("Students", "idStudent");
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Students " +
+                    "(idStudent, lastName, firstName, residence, origin, email," +
+                    //" birthDate," + // save also this field after SqlVal.SqlDate( will include the '', to better treat null values
+                    " birthPlace) " +
+                    "Values (" + codiceStudente + "," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 1]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 2]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 3]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 4]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 5]) + "'," +
+                    //"'" + SqlVal.SqlDate(StudentData[(int)StudentRow, 6]) + "'," + // save also this field after SqlVal.SqlDate( will include the ''
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 7]) + "'" +
+                    ");";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            return codiceStudente;
+        }
+
+        internal int CreateStudent(Student Student)
+        {
+            // trova una chiave da assegnare al nuovo studente
+            int idStudent = NextKey("Students", "idStudent");
+            Student.IdStudent = idStudent;
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Students " +
+                    "(idStudent,lastName,firstName,residence,origin," +
+                    "email,birthDate,birthPlace,disabled) " +
+                    "Values ('" + Student.IdStudent + "','" +
+                    SqlVal.SqlString(Student.LastName) + "','" +
+                    SqlVal.SqlString(Student.FirstName) + "','" +
+                    SqlVal.SqlString(Student.Residence) + "','" +
+                    SqlVal.SqlString(Student.Origin) + "','" +
+                    SqlVal.SqlString(Student.Email) + "'," +
+                    SqlVal.SqlDate(Student.BirthDate.ToString()) + ",'" +
+                    SqlVal.SqlString(Student.BirthPlace) + "'," +
+                    "false" +
+                    ");";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            return idStudent;
+        }
+
+        //internal void SaveStudentsOfList(List<Student> studentsList, DbConnection conn)
+        //{
+        //    foreach (Student s in studentsList)
+        //    {
+        //        SaveStudent(s, conn);
+        //    }
+        //}
+
+        internal int? SaveStudent(Student Student, DbConnection conn)
+        {
+            bool leaveConnectionOpen = true;
+            if (conn == null)
+            {
+                conn = Connect();
+                leaveConnectionOpen = false;
+            }
+            DbCommand cmd = conn.CreateCommand();
+            cmd.CommandText = "UPDATE Students " +
+                "SET" +
+                " idStudent=" + Student.IdStudent +
+                ",lastName='" + SqlVal.SqlString(Student.LastName) + "'" +
+                ",firstName='" + SqlVal.SqlString(Student.FirstName) + "'" +
+                ",residence='" + SqlVal.SqlString(Student.Residence) + "'" +
+                ",birthDate=" + SqlVal.SqlDate(Student.BirthDate.ToString()) + "" +
+                ",email='" + SqlVal.SqlString(Student.Email) + "'" +
+                //",schoolyear='" + SqlVal.SqlString(Student.SchoolYear) + "'" +
+                ",drawable='" + SqlVal.SqlBool(Student.Eligible) + "'" +
+                ",origin='" + SqlVal.SqlString(Student.Origin) + "'" +
+                ",birthPlace='" + SqlVal.SqlString(Student.BirthPlace) + "'" +
+                ",drawable=" + SqlVal.SqlBool(Student.Eligible) + "" +
+                ",disabled=" + SqlVal.SqlBool(Student.Disabled) + "" +
+                ",VFCounter=" + Student.RevengeFactorCounter + "" +
+                " WHERE idStudent = " + Student.IdStudent +
+                ";";
+            cmd.ExecuteNonQuery();
+            cmd.Dispose();
+            if (!leaveConnectionOpen)
+            {
+                conn.Close();
+                conn.Dispose();
+            }
+            return Student.IdStudent;
+        }
+
+        internal Student GetStudent(int? IdStudent)
+        {
+            Student s = new Student();
+            using (DbConnection conn = Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * " +
+                    "FROM Students " +
+                    "WHERE idStudent=" + IdStudent +
+                    ";";
+                dRead = cmd.ExecuteReader();
+                dRead.Read();
+                s = GetStudentFromRow(dRead);
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return s;
+        }
+
+        internal Student GetStudentFromRow(DbDataReader Row)
+        {
+            Student s = new Student();
+            s.IdStudent = (int)Row["IdStudent"];
+            s.LastName = SafeDb.SafeString(Row["LastName"]);
+            s.FirstName = SafeDb.SafeString(Row["FirstName"]);
+            s.Residence = SafeDb.SafeString(Row["Residence"]);
+            s.Origin = SafeDb.SafeString(Row["Origin"]);
+            s.Email = SafeDb.SafeString(Row["Email"]);
+            if (!(Row["birthDate"] is DBNull))
+                s.BirthDate = SafeDb.SafeDateTime(Row["birthDate"]);
+            s.BirthPlace = SafeDb.SafeString(Row["birthPlace"]);
+            s.Eligible = SafeDb.SafeBool(Row["drawable"]);
+            s.Disabled = SafeDb.SafeBool(Row["disabled"]);
+            s.RevengeFactorCounter = SafeDb.SafeInt(Row["VFCounter"]);
+
+            return s;
+        }
+
+        internal DataTable GetStudentsSameName(string LastName, string FirstName)
+        {
+            DataTable t;
+            using (DbConnection conn = Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT Students.IdStudent, Students.lastName, Students.firstName," +
+                    " Classes.abbreviation, Classes.idSchoolYear" +
+                    " FROM Students" +
+                    " LEFT JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent " +
+                    " LEFT JOIN Classes ON Classes.idClass = Classes_Students.idClass " +
+                    " WHERE Students.lastName LIKE '%" + LastName + "%'" +
+                    " AND Students.firstName LIKE '%" + FirstName + "%'" +
+                    ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetStudentsSameName");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dSet.Dispose();
+                dAdapt.Dispose();
+            }
+            return t;
+        }
+
+        internal DataTable FindStudentsLike(string LastName, string FirstName)
+        {
+            DataTable t;
+            using (DbConnection conn = Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT Students.IdStudent, Students.lastName, Students.firstName," +
+                    " Classes.abbreviation, Classes.idSchoolYear" +
+                    " FROM Students" +
+                    " LEFT JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent " +
+                    " LEFT JOIN Classes ON Classes.idClass = Classes_Students.idClass ";
+                if (LastName != "" && LastName != null)
+                {
+                    query += "WHERE Students.lastName LIKE '%" + LastName + "%'";
+                    if (FirstName != "" && FirstName != null)
+                    {
+                        query += " AND Students.firstName LIKE '%" + FirstName + "%'";
+                    }
+                }
+                else
+                {
+                    if (FirstName != "" && FirstName != null)
+                    {
+                        query += " WHERE Students.firstName LIKE '%" + FirstName + "%'";
+                    }
+                }
+                query += ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetStudentsSameName");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dSet.Dispose();
+                dAdapt.Dispose();
+            }
+            return t;
+        }
+
+        internal void PutStudentInClass(int? IdStudent, int? IdClass)
+        {
+            using (DbConnection conn = Connect())
+            {
+                // add student to the class
+                DbCommand cmd = conn.CreateCommand();
+                // check if already present
+                cmd.CommandText = "SELECT IdStudent FROM Classes_Students " +
+                    "WHERE idClass=" + IdClass + " AND idStudent=" + IdStudent + "" +
+                ";";
+                if (cmd.ExecuteScalar() == null)
+                {
+                    // add student to the class
+                    cmd.CommandText = "INSERT INTO Classes_Students " +
+                    "(idClass, idStudent) " +
+                    "Values ('" + IdClass + "'," + IdStudent + "" +
+                    ");";
+                }
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="IdClass">Id of the class to be searched</param>
+        /// <param name="conn">Connection already open on a database different from standard. 
+        /// If not null this connection is left open</param>
+        /// <returns>List of the </returns>
+        internal List<Student> GetStudentsOfClass(int? IdClass, DbConnection conn)
+        {
+            List<Student> l = new List<Student>();
+            bool leaveConnectionOpen = true;
+
+            if (conn == null)
+            {
+                conn = Connect();
+                leaveConnectionOpen = false;
+            }
+            DbDataReader dRead;
+            DbCommand cmd = conn.CreateCommand();
+            string query = "SELECT Students.*" +
+                " FROM Students" +
+                " JOIN Classes_Students ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass=" + IdClass +
+            ";";
+            cmd.CommandText = query;
+            dRead = cmd.ExecuteReader();
+
+            while (dRead.Read())
+            {
+                Student s = GetStudentFromRow(dRead);
+                l.Add(s);
+            }
+            if (!leaveConnectionOpen)
+            {
+                conn.Close();
+                conn.Dispose();
+            }
+            return l;
+        }
+
+        internal List<Student> GetStudentsOfClassList(string Scuola, string Anno,
+            string SiglaClasse, bool IncludeNonActiveStudents)
+        {
+            DbDataReader dRead;
+            DbCommand cmd;
+            List<Student> ls = new List<Student>();
+            using (DbConnection conn = Connect())
+            {
+                string query = "SELECT registerNumber, Classes.idSchoolYear, " +
+                               "Classes.abbreviation, Classes.idClass, Classes.idSchool, " +
+                               "Students.*" +
+                " FROM Students" +
+                " JOIN Classes_Students ON Students.idStudent=Classes_Students.idStudent" +
+                " JOIN Classes ON Classes.idClass=Classes_Students.idClass" +
+                " WHERE Classes.idSchoolYear = '" + SqlVal.SqlString(Anno) + "'" +
+                " AND Classes.abbreviation = '" + SqlVal.SqlString(SiglaClasse) + "'";
+                if (!IncludeNonActiveStudents)
+                    query += " AND (Students.disabled = 0 OR Students.disabled IS NULL)";
+                if (Scuola != null && Scuola != "")
+                    query += " AND Classes.idSchool='" + Scuola + "'";
+                query += " ORDER BY Students.LastName, Students.FirstName";
+                query += ";";
+                cmd = conn.CreateCommand();
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    Student s = GetStudentFromRow(dRead);
+                    s.Class = (string)dRead["abbreviation"];
+                    s.IdClass = (int)dRead["idClass"];
+                    ls.Add(s);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return ls;
+        }
+
+        internal List<int> GetIdStudentsNonGraded(Class Class,
+            GradeType GradeType, SchoolSubject SchoolSubject)
+        {
+            List<int> keys = new List<int>();
+
+            DbDataReader dRead;
+            DbCommand cmd;
+            using (DbConnection conn = Connect())
+            {
+                string query = "SELECT Classes_Students.idStudent" +
+                " FROM Classes_Students" +
+                " WHERE Classes_Students.idClass=" + Class.IdClass +
+                " AND Classes_Students.idStudent NOT IN" +
+                "(" +
+                "SELECT DISTINCT Classes_Students.idStudent" +
+                " FROM Classes_Students" +
+                " LEFT JOIN Grades ON Classes_Students.idStudent = Grades.idStudent" +
+                " WHERE Classes_Students.idClass=" + Class.IdClass +
+                " AND Grades.idSchoolSubject='" + SchoolSubject.IdSchoolSubject + "'" +
+                " AND Grades.idGradeType='" + GradeType.IdGradeType + "'" +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                ")" +
+                ";";
+                cmd = conn.CreateCommand();
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    keys.Add((int)SafeDb.SafeInt(dRead["idStudent"]));
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return keys;
+        }
+
+
+    }
+}
